# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Restaurant API",
        "contact": {
            "name": "Hojune Yoo",
            "email": "soy0803@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders",
                "operationId": "GetAllOrders",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Save an order",
                "operationId": "SaveOrder",
                "parameters": [
                    {
                        "name": "special_request",
                        "in": "formData",
                        "description": "message context",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "option_selected",
                        "in": "formData",
                        "description": "option description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "visit_id",
                        "in": "formData",
                        "description": "Visit Id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "item_id",
                        "in": "formData",
                        "description": "Item Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/orders/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get an Order",
                "operationId": "GetOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "DeleteOrder",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}